netcdf_to_tplot: addressed desired fixes/updates/changes
